Closure based URC handeling, fixed closing sockets, DNS spin error fix plus various other fixes

+++ ublox-short-range/src/command/data_mode/urc.rs
@@ -33,5 +33,5 @@ pub struct PeerConnected {
 #[derive(Debug, PartialEq, Clone, AtatResp)]
 pub struct PeerDisconnected {
     #[at_arg(position = 0)]
-    pub handle: u32,
+    pub handle: usize,
 }
